scripts/markdown-to-pdf/pyproject.toml
[project]
name = "markdown-to-pdf"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
  "markdown-it-py>=3.0.0",
  "weasyprint>=68.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
